chore: Split parse from embed and experiment
- compare text-embedding-3-large vs text-embedding-3-small
- compare embedding CDKTF Aws Markdown docs Summary + Arguments sections
vs embedding docs summary only
- compare GraphRAG query with full Cfn TS Interface declarations vs only
using the JSDocs section of the Cfn TS Class declaration
